Version 1 of todo-list

+++ server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 const date = require(__dirname + "/date.js");
 
 const app = express();
@@ -11,13 +12,24 @@ app.use(bodyParser.urlencoded({extended : true}));
 app.use(express.static("public"));
 
 //Connecting to the database and creating the table schema
-mongoose.connect("mongodb://127.0.0.1:27017/demoDb").then(() => {
+mongoose.connect("mongodb+srv://rajadarsh268:adarsh_2608@cluster0.t3cihiw.mongodb.net/demoDb").then(() => {
     console.log("Connected to MongoDB!")
 }).catch((err) => {
     console.log(err)
 })
+
+//creating schema for homePage todo-list
 const taskSchema = new mongoose.Schema({task : String});
 const task = new mongoose.model("task", taskSchema);
+
+//Creating another schema to store other kinds of task
+const otherTaskSchema = new mongoose.Schema({
+  categoryName : String,
+  taskList : [taskSchema]
+});
+
+const otherTask = new mongoose.model("otherTask" , otherTaskSchema);
+
 
 // //Create opertion
 // const adder = async() => {
@@ -49,6 +61,11 @@ const task = new mongoose.model("task", taskSchema);
 // }
 // reader();
 
+//DEFAULT ITEMS TO DISPLAY
+const t1 = new task({task : "Welcome to to-do list App!"});
+const t2 = new task({task : "Hit the + button to add task"});
+const t3 = new task({task : "<-- to delete task"});
+
 //HOME ROUTE
 app.get("/", async (req, res) =>{
 
@@ -59,10 +76,6 @@ app.get("/", async (req, res) =>{
     if(todoEntries.length == 0){
       const adderbasic = async() => {
         try{
-          const t1 = new task({task : "Welcome to to-do list App!"});
-          const t2 = new task({task : "Hit the + button to add task"});
-          const t3 = new task({task : "<-- to delete task"});
-        
           const data = await task.insertMany([t1, t2, t3]);
           console.log(data);
         }catch(err){
@@ -85,15 +98,60 @@ app.get("/", async (req, res) =>{
   }
 })
 
-//ROUTE FOR ADDING ITEMS
+//OTHER TASKS ROUTE
+app.get("/:customRoute", async(req, res) => {
+
+  try {
+    const customRouteName = _.capitalize(req.params.customRoute);
+    const customRouteData = await otherTask.findOne({categoryName : customRouteName});
+
+    if(!customRouteData){
+      //if not present then show the intro list
+      const adderOthertask = async() => {
+        try{
+          const data = await otherTask.create({
+            categoryName: customRouteName,
+            taskList : [t1, t2, t3]
+          });
+          console.log(data);
+        }catch(err){
+          console.log(err.message);
+        }
+
+      }
+      adderOthertask();
+      res.redirect("/" + customRouteName);       
+    }else{
+      //display the current list
+      console.log(customRouteData);
+      res.render("list.ejs" , {listTitle: customRouteData.categoryName, newListItems: customRouteData.taskList});
+    }
+
+  }catch (err) {
+    console.log(err.message);
+  }
+});
+
+//ROUTE FOR ADDING ITEMS 
 app.post("/", async(req, res) =>{
   /*First getting the content from the post request, adding it to the database and redirecting it 
   it to the home route*/
   try{
     const taskToAdd = req.body.newItem;
+    const locationofAdd = req.body.list;
+
     const nextTask = new task({task: taskToAdd});
-    await nextTask.save();
-    res.redirect("/");
+    if(locationofAdd === "Today"){
+      await nextTask.save();
+      res.redirect("/");
+    }else{
+      const customRouteData = await otherTask.findOne({categoryName : locationofAdd});
+      customRouteData.taskList.push(nextTask);
+      await customRouteData.save();
+      res.redirect("/" + locationofAdd);
+    }
+    
+
   }catch(err){
     console.log(err.message);
   }
@@ -105,9 +163,21 @@ app.post("/delete", async(req,res) =>{
   /* getting the item's ID which needs to be deleted via EJS file, deleting 
   it from the database and finally redirecting it the the home route*/
   try {
+    const deleteLocation = req.body.listTitle;
     const deleteId = req.body.checkbox;
-    await task.deleteOne({_id: deleteId});
-    res.redirect("/");
+
+    if(deleteLocation === "Today"){
+      await task.deleteOne({_id: deleteId});
+      res.redirect("/");
+    }else{
+      const customRouteData = await otherTask.findOneAndUpdate(
+        {categoryName : deleteLocation},
+        {$pull : {taskList : {_id: deleteId}}}
+      );
+      res.redirect("/" + deleteLocation);
+
+    }
+
 
   }catch (err) {
     console.log(err.message);
